backend/requirements.txt
Flask==2.2.5
Flask-PyMongo==2.3.0
pymongo==3.12.0
cryptography==3.4.7
Python-dotenv==0.19.0
Flask-CORS==3.0.10
Werkzeug==2.2.2
